Created new Directory for Images
As it currently stands, there are several folders/directories for each individual blog post (example: PWA folder for the 6 images in that blog). I know I need permission to upload files, and I think it would be a long process to create a folder every time I upload a single image for an article. I suggest creating this folder so I can keep the images together in one place. If this isn't efficient please advise.